Refactor character movement logic and types
This commit primarily changes logic related to character movement in Pac-Man game and refactors some relevant types for better clarity and optimization. Internal 'position' state of the characters has been updated to use a new, more descriptive type 'DirectionalPosition'. This change helps clarify the meaning of the state, and adds explicit direction information to each position.

Change replaces methods in some of the classes with property get methods, allowing a cleaner and more intuitive access to the relevant attributes. The algorithm calculating possible moves has been adjusted to work with the new 'DirectionalPosition' type. As a result, tests related to possible move calculation, as well as some additional logic that processes player actions were updated to reflect these changes. Changes enhances clarity and readability of the code, making it easier to understand and maintain.

Boolean properties in 'IWebSocket' interface were also refactored to get rid of the clunky method calls - now they are getters.

NOTE: The change should not impact the game logic or sequence of events in the game flow, it's a purely internal code cleanup and optimization.